🐷

+++ week12/pE.c
@@ -13,5 +13,5 @@ int main(void) {
 
 double calculate_score(int CH, int ENG, int *MATH, int SCI, int SS, int *TOTAL) {
     *TOTAL = CH + ENG + *MATH + SCI + SS;
-    return *TOTAL / (double)5;
+    return *TOTAL / 5.0;
 }
